style(circleci): move api endpoints to seperate modules
organize the structure of the library by creating seperate modules
