+++ src/api/actions.ts
@@ -1,0 +1,51 @@
+import {
+  GitInfo,
+  BuildActionResponse,
+  FullRequest,
+  TriggerBuildResponse
+} from "../types";
+import { BuildAction, createVcsUrl } from "../circleci";
+import { client } from "../client";
+
+/**
+ * Commit a build action, returns a summary of the new build.
+ *
+ * Retry a build
+ * @see https://circleci.com/docs/api/v1-reference/#retry-build
+ * @example POST - /project/:vcs-type/:username/:project/:build_num/retry
+ *
+ * Cancel a build
+ * @see https://circleci.com/docs/api/v1-reference/#cancel-build
+ * @example POST - /project/:vcs-type/:username/:project/:build_num/cancel
+ *
+ * @param token - CircleCI API token
+ * @param vcs - Project's git information that you'd like to retry
+ * @param buildNumber - Target build number to retry
+ * @param action - Action to perform on the build
+ */
+export function buildActions(
+  token: string,
+  vcs: GitInfo,
+  buildNumber: number,
+  action: BuildAction
+): Promise<BuildActionResponse> {
+  const url = `${createVcsUrl(vcs)}/${buildNumber}/${action}`;
+  return client(token).post(url);
+}
+
+/**
+ * Triggers a new build, returns a summary of the build.
+ * @see https://circleci.com/docs/api/v1-reference/#new-build
+ * @example /project/:vcs-type/:username/:project
+ *
+ * Triggers a new build, returns a summary of the build.
+ * @see https://circleci.com/docs/api/v1-reference/#new-build-branch
+ * @example /project/:vcs-type/:username/:project/tree/:branch
+ */
+export function triggerNewBuild(
+  token: string,
+  { vcs, options: { branch = "", newBuildOptions = {} } = {} }: FullRequest
+): Promise<TriggerBuildResponse> {
+  const url = `${createVcsUrl(vcs)}${branch ? `/tree/${branch}` : ""}`;
+  return client(token).post<TriggerBuildResponse>(url, newBuildOptions);
+}
