+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./types/api";
 export * from "./types/lib";
+export * from "./api";
 export * from "./circleci";
 export { getGitType } from "./util";
